chatbot-ai/requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
sentence-transformers==2.2.2
chromadb==0.4.22
numpy==1.24.3
pandas==2.0.3

openai==1.3.0  # Optional, nếu dùng OpenAI API
tiktoken==0.5

# FIX AI stack
huggingface_hub==0.14.1
transformers==4.30.2
tokenizers==0.13.3